Tested basic logging function
More testing with custom modes coming.
