+++ customLogger.py
@@ -1,3 +1,4 @@
+import sys
 import logging
 
 # Code from https://stackoverflow.com/a/28636024/3769258
@@ -15,3 +16,43 @@ class LevelFormatter(logging.Formatter):
             return self._level_formatters[record.levelno].format(record)
 
         return super(LevelFormatter, self).format(record)
+
+def gen_logging(filename, verbose=False, verbose_flag = None):
+
+    root = logging.getLogger(__name__)
+
+    if filename != "":
+        log_f = logging.FileHandler(filename + ".log")
+    ch = logging.StreamHandler(sys.stdout)
+
+    root.setLevel(logging.INFO)
+    ch.setLevel(logging.INFO)
+
+    if verbose:
+        if verbose_flag.lower() == 'debug':
+            root.setLevel(logging.DEBUG)
+            ch.setLevel(logging.DEBUG)
+        elif verbose_flag.lower() == 'info':
+            root.setLevel(logging.INFO)
+            ch.setLevel(logging.INFO)
+        elif verbose_flag.lower() == 'warning':
+            root.setLevel(logging.WARNING)
+            ch.setLevel(logging.WARNING)
+        elif verbose_flag.lower() == 'error':
+            root.setLevel(logging.ERROR)
+            ch.setLevel(logging.ERROR)
+        elif verbose_flag.lower() == 'critical':
+            root.setLevel(logging.CRITICAL)
+            ch.setLevel(logging.CRITICAL)
+
+    formatter = LevelFormatter(fmt="[%(asctime)s] %(levelname)s: %(message)s", level_fmts={logging.INFO: "%(message)s"})
+
+    if filename != "":
+        log_f.setFormatter(formatter)
+    ch.setFormatter(formatter)
+
+    if filename != "":
+        root.addHandler(log_f)
+    root.addHandler(ch)
+
+    return root
