+++ simulation.py
@@ -3,7 +3,7 @@ import numpy as np
 from person import Person
 from epidemic import Epidemic
 from contact import ContactNwk
-from levelFormatter import LevelFormatter
+import customLogger
 import write
 
 import random
